+++ Entities/IncomeSource.cs
@@ -9,9 +9,9 @@ namespace BudgetBucketsAPI.Entities
 		[ForeignKey("Income")]
 		public int IncomeId {get; set;}
 		public string Name {get; set;} = String.Empty;
-		public int AmountPlanned {get; set;}
-		public int AmountReceived {get; set;}
-		public int AmountRemaining {get; set;}
+		public decimal AmountPlanned {get; set;}
+		public decimal AmountReceived {get; set;}
+		public decimal AmountRemaining {get; set;}
 		public string Note {get; set;} = String.Empty;
 	}
 }
